requirements.txt
Pillow==8.4.0
Pillow==9.2.0
Pillow==9.5.0
streamlit==1.23.1
torch==1.12.1
torchvision==0.13.1
